+++ go/tasks/plugins/k8s/pod/container.go
@@ -1,0 +1,28 @@
+package pod
+
+import (
+	"context"
+
+	"github.com/flyteorg/flyteidl/gen/pb-go/flyteidl/core"
+
+	pluginsCore "github.com/flyteorg/flyteplugins/go/tasks/pluginmachinery/core"
+	"github.com/flyteorg/flyteplugins/go/tasks/pluginmachinery/flytek8s"
+
+	v1 "k8s.io/api/core/v1"
+)
+
+type containerPodBuilder struct {
+}
+
+func (containerPodBuilder) BuildPodSpec(ctx context.Context, task *core.TaskTemplate, taskCtx pluginsCore.TaskExecutionContext) (*v1.PodSpec, error) {
+	podSpec, err := flytek8s.ToK8sPodSpec(ctx, taskCtx)
+	if err != nil {
+		return nil, err
+	}
+
+	return podSpec, nil
+}
+
+func (containerPodBuilder) UpdatePodMetadata(ctx context.Context, pod *v1.Pod, task *core.TaskTemplate, taskCtx pluginsCore.TaskExecutionContext) error {
+	return nil
+}
